[update] structure directories (ex. src/js/main.js)
I structured directory. Moving all js files to src/js and changing
package.json
