+++ index.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="UTF-8" />
 		<title> TODO hwata</title>
-		<link rel="stylesheet" type="text/css" href="style.css">
+		<link rel="stylesheet" type="text/css" href="./src/css/style.css">
 	</head>
 	<body>
 
@@ -21,6 +21,6 @@
 			<ul class="doneArea">
 			</ul>
 		</section>
-		<script src="todo.js"></script>
+		<script type="module" src="./src/js/todo.js"></script>
 	</body>
 </html>
