--- memo.sh
@@ -1,1 +1,0 @@
-npx serve ./src
