+++ src/js/main.js
@@ -1,58 +1,13 @@
 "use strict"
-/**
- * @description 
- * Enums for todo status <br>
- * <strong>why is status defined as Object?</strong><br>
- * 1. change and add status easily by using Object <br>
- * 2. search easily on editor by using StatusEnum word because this code includes many todoXxx value <br>
- * 3. prevent spell mistake <br>
- * @typedef {Object} StatusEnum
- * @property {string} TODO - Represents a todo status
- * @property {string} DONE - Represents a done status
- * @property {string} DELETED - Represents a deleted status
- * @readonly
- */
-const StatusEnum = Object.freeze({
-    TODO: "todo",
-    DONE: "done",
-    DELETED: "deleted"
-});
-
-/**
- * @description
- * check if todoItem.status is done or not
- * @param {Todo | undefined} todoItem checked todo object
- * @return {boolean}
-*/
-function isStatusDone(todoItem) {
-    return todoItem?.status === StatusEnum.DONE;
-}
-
-/**
- * @description
- * check if todoItem.status is deleted or not
- * @param {Todo | undefined} todoItem checked todo object
- * @return {boolean}
-*/
-function isStatusDeleted(todoItem) {
-    return todoItem?.status === StatusEnum.DELETED;
-}
-/**
- * @description
- * check if todoItem.status is todo or not
- * @param {Todo | undefined} todoItem checked todo object
- * @return {boolean}
-*/
-function isStatusTodo(todoItem) {
-    return todoItem?.status === StatusEnum.TODO;
-}
+import { statusEnum } from "./statusEnum.js";
+import { isStatusDone, isStatusDeleted, isStatusTodo } from './utils.js';
 
 /**
  * todos has todo's content and status value
  * @typedef {Object} Todo - todo object
  * @property {string} id - Todo id (uuid). It is used when change indivisual todo item.
  * @property {string} content - The description of the todo item.
- * @property {StatusEnum} status - todo status that defines statusEnum.
+ * @property {statusEnum} status - todo status that defines statusEnum.
 **/
 /** @type {Todo[]} */
 const todos = [];
@@ -61,7 +16,7 @@ const todos = [];
  * @returns {void}
  * @param {string} id Todo id (uuid).
  * @param {string} inputContent todo content field.
- * @param {StatusEnum} inputStatus todo status filld.
+ * @param {statusEnum} inputStatus todo status filld.
  */
 function appendTodos(inputContent, inputStatus) {
     const newTask = {
@@ -110,7 +65,7 @@ function createTodo() {
 
     //add todo to todos
     const todoArea = document.querySelector(".todoArea");
-    appendTodos(input.value, StatusEnum.TODO);
+    appendTodos(input.value, statusEnum.TODO);
     
     // create elms
     const todoItemWrapper = createElementWrapper("div", "todoItemWrapper", null);
@@ -122,10 +77,10 @@ function createTodo() {
     const todoItem = createElementWrapper("li", "todoItem", null);
     const todoContent = createElementWrapper("span", "todoContent", input.value);
 
-    const doneButton = createElementWrapper("button", "doneButton", StatusEnum.DONE);
+    const doneButton = createElementWrapper("button", "doneButton", statusEnum.DONE);
     doneButton.addEventListener("click", () => doneTodo(todoItemWrapper.id));
 
-    const deleteButton = createElementWrapper("button", "deleteButton", StatusEnum.DELETED);
+    const deleteButton = createElementWrapper("button", "deleteButton", statusEnum.DELETED);
     deleteButton.addEventListener("click", () => deleteTodo(todoItemWrapper.id));
        
     // add todo elms to todoArea by using DOM handle
@@ -175,7 +130,7 @@ function deleteTodo(deleteId) {
 function isChangeStatusDone(todoItem) {
     const hasStatus = todoItem.hasOwnProperty("status");
     if (hasStatus) {
-        todoItem.status = StatusEnum.DONE;
+        todoItem.status = statusEnum.DONE;
     }
 
     return hasStatus;
